+++ test_unstructured/partition/test_auto.py
@@ -41,7 +41,7 @@ odt_not_supported = "odt" not in pypandoc.get_pandoc_formats()[0]
 
 def test_auto_partition_email_from_filename():
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email.eml")
-    elements = partition(filename=filename)
+    elements = partition(filename=filename, strategy="hi_res")
     assert len(elements) > 0
     assert elements == EXPECTED_EMAIL_OUTPUT
     assert elements[0].metadata.filename == os.path.basename(filename)
@@ -51,7 +51,7 @@ def test_auto_partition_email_from_filename():
 def test_auto_partition_email_from_file():
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email.eml")
     with open(filename) as f:
-        elements = partition(file=f)
+        elements = partition(file=f, strategy="hi_res")
     assert len(elements) > 0
     assert elements == EXPECTED_EMAIL_OUTPUT
 
@@ -59,7 +59,7 @@ def test_auto_partition_email_from_file():
 def test_auto_partition_email_from_file_rb():
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email.eml")
     with open(filename, "rb") as f:
-        elements = partition(file=f)
+        elements = partition(file=f, strategy="hi_res")
     assert len(elements) > 0
     assert elements == EXPECTED_EMAIL_OUTPUT
 
@@ -102,7 +102,7 @@ def test_auto_partition_docx_with_filename(mock_docx_document, expected_docx_ele
     filename = os.path.join(tmpdir.dirname, "mock_document.docx")
     mock_docx_document.save(filename)
 
-    elements = partition(filename=filename)
+    elements = partition(filename=filename, strategy="hi_res")
     assert elements == expected_docx_elements
     assert elements[0].metadata.filename == os.path.basename(filename)
 
@@ -112,7 +112,7 @@ def test_auto_partition_docx_with_file(mock_docx_document, expected_docx_element
     mock_docx_document.save(filename)
 
     with open(filename, "rb") as f:
-        elements = partition(file=f)
+        elements = partition(file=f, strategy="hi_res")
     assert elements == expected_docx_elements
 
 
@@ -136,6 +136,7 @@ def test_auto_partition_doc_with_filename(
         filename=doc_filename,
         file_filename=file_filename,
         content_type=content_type,
+        strategy="hi_res",
     )
     assert elements == expected_docx_elements
     assert elements[0].metadata.filename == "mock_document.doc"
@@ -152,7 +153,7 @@ def test_auto_partition_doc_with_file(mock_docx_document, expected_docx_elements
     convert_office_doc(docx_filename, tmpdir.dirname, "doc")
 
     with open(doc_filename, "rb") as f:
-        elements = partition(file=f)
+        elements = partition(file=f, strategy="hi_res")
     assert elements == expected_docx_elements
 
 
@@ -163,7 +164,12 @@ def test_auto_partition_doc_with_file(mock_docx_document, expected_docx_elements
 def test_auto_partition_html_from_filename(pass_file_filename, content_type):
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "example-10k.html")
     file_filename = filename if pass_file_filename else None
-    elements = partition(filename=filename, file_filename=file_filename, content_type=content_type)
+    elements = partition(
+        filename=filename,
+        file_filename=file_filename,
+        content_type=content_type,
+        strategy="hi_res",
+    )
     assert len(elements) > 0
     assert elements[0].metadata.filename == os.path.basename(filename)
     assert elements[0].metadata.file_directory == os.path.split(filename)[0]
@@ -177,14 +183,19 @@ def test_auto_partition_html_from_file(pass_file_filename, content_type):
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-html.html")
     file_filename = filename if pass_file_filename else None
     with open(filename) as f:
-        elements = partition(file=f, file_filename=file_filename, content_type=content_type)
+        elements = partition(
+            file=f,
+            file_filename=file_filename,
+            content_type=content_type,
+            strategy="hi_res",
+        )
     assert len(elements) > 0
 
 
 def test_auto_partition_html_from_file_rb():
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-html.html")
     with open(filename, "rb") as f:
-        elements = partition(file=f)
+        elements = partition(file=f, strategy="hi_res")
     assert len(elements) > 0
 
 
@@ -200,7 +211,7 @@ def test_auto_partition_json_from_filename():
     )
     with open(filename) as json_f:
         json_data = json.load(json_f)
-    json_elems = json.loads(elements_to_json(partition(filename=filename)))
+    json_elems = json.loads(elements_to_json(partition(filename=filename, strategy="hi_res")))
     for elem in json_elems:
         # coordinates are always in the element data structures, even if None
         elem.pop("coordinates")
@@ -226,7 +237,7 @@ def test_auto_partition_json_from_file():
     with open(filename) as json_f:
         json_data = json.load(json_f)
     with open(filename, encoding="utf-8") as partition_f:
-        json_elems = json.loads(elements_to_json(partition(file=partition_f)))
+        json_elems = json.loads(elements_to_json(partition(file=partition_f, strategy="hi_res")))
     for elem in json_elems:
         # coordinates are always in the element data structures, even if None
         elem.pop("coordinates")
@@ -245,7 +256,7 @@ EXPECTED_TEXT_OUTPUT = [
 
 def test_auto_partition_text_from_filename():
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-text.txt")
-    elements = partition(filename=filename)
+    elements = partition(filename=filename, strategy="hi_res")
     assert len(elements) > 0
     assert elements == EXPECTED_TEXT_OUTPUT
     assert elements[0].metadata.filename == os.path.basename(filename)
@@ -255,7 +266,7 @@ def test_auto_partition_text_from_filename():
 def test_auto_partition_text_from_file():
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-text.txt")
     with open(filename) as f:
-        elements = partition(file=f)
+        elements = partition(file=f, strategy="hi_res")
     assert len(elements) > 0
     assert elements == EXPECTED_TEXT_OUTPUT
 
@@ -268,7 +279,12 @@ def test_auto_partition_pdf_from_filename(pass_file_filename, content_type):
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "layout-parser-paper-fast.pdf")
     file_filename = filename if pass_file_filename else None
 
-    elements = partition(filename=filename, file_filename=file_filename, content_type=content_type)
+    elements = partition(
+        filename=filename,
+        file_filename=file_filename,
+        content_type=content_type,
+        strategy="hi_res",
+    )
 
     assert isinstance(elements[0], Title)
     assert elements[0].text.startswith("LayoutParser")
@@ -285,7 +301,7 @@ def test_auto_partition_pdf_uses_table_extraction():
     with patch(
         "unstructured_inference.inference.layout.process_file_with_model",
     ) as mock_process_file_with_model:
-        partition(filename, pdf_infer_table_structure=True)
+        partition(filename, pdf_infer_table_structure=True, strategy="hi_res")
         assert mock_process_file_with_model.call_args[1]["extract_tables"]
 
 
@@ -317,7 +333,12 @@ def test_auto_partition_pdf_from_file(pass_file_filename, content_type):
     file_filename = filename if pass_file_filename else None
 
     with open(filename, "rb") as f:
-        elements = partition(file=f, file_filename=file_filename, content_type=content_type)
+        elements = partition(
+            file=f,
+            file_filename=file_filename,
+            content_type=content_type,
+            strategy="hi_res",
+        )
 
     assert isinstance(elements[0], Title)
     assert elements[0].text.startswith("LayoutParser")
@@ -334,7 +355,7 @@ def test_partition_pdf_doesnt_raise_warning():
     #      #additional-use-cases-of-warnings-in-tests
     with warnings.catch_warnings():
         warnings.simplefilter("error")
-        partition(filename=filename)
+        partition(filename=filename, strategy="hi_res")
 
 
 @pytest.mark.parametrize(
@@ -344,7 +365,12 @@ def test_partition_pdf_doesnt_raise_warning():
 def test_auto_partition_jpg(pass_file_filename, content_type):
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "layout-parser-paper-fast.jpg")
     file_filename = filename if pass_file_filename else None
-    elements = partition(filename=filename, file_filename=file_filename, content_type=content_type)
+    elements = partition(
+        filename=filename,
+        file_filename=file_filename,
+        content_type=content_type,
+        strategy="hi_res",
+    )
     assert len(elements) > 0
 
 
@@ -356,14 +382,19 @@ def test_auto_partition_jpg_from_file(pass_file_filename, content_type):
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "layout-parser-paper-fast.jpg")
     file_filename = filename if pass_file_filename else None
     with open(filename, "rb") as f:
-        elements = partition(file=f, file_filename=file_filename, content_type=content_type)
+        elements = partition(
+            file=f,
+            file_filename=file_filename,
+            content_type=content_type,
+            strategy="hi_res",
+        )
     assert len(elements) > 0
 
 
 def test_auto_partition_raises_with_bad_type(monkeypatch):
     monkeypatch.setattr(auto, "detect_filetype", lambda *args, **kwargs: None)
     with pytest.raises(ValueError):
-        partition(filename="made-up.fake")
+        partition(filename="made-up.fake", strategy="hi_res")
 
 
 EXPECTED_PPTX_OUTPUT = [
@@ -378,7 +409,7 @@ EXPECTED_PPTX_OUTPUT = [
 
 def test_auto_partition_pptx_from_filename():
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-power-point.pptx")
-    elements = partition(filename=filename)
+    elements = partition(filename=filename, strategy="hi_res")
     assert elements == EXPECTED_PPTX_OUTPUT
     assert elements[0].metadata.filename == os.path.basename(filename)
     assert elements[0].metadata.file_directory == os.path.split(filename)[0]
@@ -387,7 +418,7 @@ def test_auto_partition_pptx_from_filename():
 @pytest.mark.skipif(is_in_docker, reason="Skipping this test in Docker container")
 def test_auto_partition_ppt_from_filename():
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-power-point.ppt")
-    elements = partition(filename=filename)
+    elements = partition(filename=filename, strategy="hi_res")
     assert elements == EXPECTED_PPTX_OUTPUT
     assert elements[0].metadata.filename == os.path.basename(filename)
     assert elements[0].metadata.file_directory == os.path.split(filename)[0]
@@ -395,14 +426,14 @@ def test_auto_partition_ppt_from_filename():
 
 def test_auto_with_page_breaks():
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "layout-parser-paper-fast.pdf")
-    elements = partition(filename=filename, include_page_breaks=True)
+    elements = partition(filename=filename, include_page_breaks=True, strategy="hi_res")
     assert PageBreak() in elements
 
 
 @pytest.mark.skipif(is_in_docker, reason="Skipping this test in Docker container")
 def test_auto_partition_epub_from_filename():
     filename = os.path.join(DIRECTORY, "..", "..", "example-docs", "winter-sports.epub")
-    elements = partition(filename=filename)
+    elements = partition(filename=filename, strategy="hi_res")
     assert len(elements) > 0
     assert elements[0].text.startswith("The Project Gutenberg eBook of Winter Sports")
 
@@ -411,7 +442,7 @@ def test_auto_partition_epub_from_filename():
 def test_auto_partition_epub_from_file():
     filename = os.path.join(DIRECTORY, "..", "..", "example-docs", "winter-sports.epub")
     with open(filename, "rb") as f:
-        elements = partition(file=f)
+        elements = partition(file=f, strategy="hi_res")
     assert len(elements) > 0
     assert elements[0].text.startswith("The Project Gutenberg eBook of Winter Sports")
 
@@ -426,7 +457,7 @@ EXPECTED_MSG_OUTPUT = [
 
 def test_auto_partition_msg_from_filename():
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email.msg")
-    elements = partition(filename=filename)
+    elements = partition(filename=filename, strategy="hi_res")
     assert elements == EXPECTED_MSG_OUTPUT
 
 
@@ -434,20 +465,20 @@ def test_auto_partition_msg_from_filename():
 @pytest.mark.skipif(rtf_not_supported, reason="RTF not supported in this version of pypandoc.")
 def test_auto_partition_rtf_from_filename():
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-doc.rtf")
-    elements = partition(filename=filename)
+    elements = partition(filename=filename, strategy="hi_res")
     assert elements[0] == Title("My First Heading")
 
 
 def test_auto_partition_from_url():
     url = "https://raw.githubusercontent.com/Unstructured-IO/unstructured/main/LICENSE.md"
-    elements = partition(url=url, content_type="text/plain")
+    elements = partition(url=url, content_type="text/plain", strategy="hi_res")
     assert elements[0] == Title("Apache License")
     assert elements[0].metadata.url == url
 
 
 def test_partition_md_works_with_embedded_html():
     url = "https://raw.githubusercontent.com/Unstructured-IO/unstructured/main/README.md"
-    elements = partition(url=url, content_type="text/markdown")
+    elements = partition(url=url, content_type="text/markdown", strategy="hi_res")
     elements[0].text
     unstructured_found = False
     for element in elements:
@@ -459,13 +490,13 @@ def test_partition_md_works_with_embedded_html():
 
 def test_auto_partition_warns_if_header_set_and_not_url(caplog):
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email.eml")
-    partition(filename=filename, headers={"Accept": "application/pdf"})
+    partition(filename=filename, headers={"Accept": "application/pdf"}, strategy="hi_res")
     assert caplog.records[0].levelname == "WARNING"
 
 
 def test_auto_partition_works_with_unstructured_jsons():
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "spring-weather.html.json")
-    elements = partition(filename=filename)
+    elements = partition(filename=filename, strategy="hi_res")
     assert elements[0].text == "News Around NOAA"
 
 
@@ -473,7 +504,7 @@ def test_auto_partition_works_with_unstructured_jsons_from_file():
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "spring-weather.html.json")
 
     with open(filename, "rb") as f:
-        elements = partition(file=f)
+        elements = partition(file=f, strategy="hi_res")
     assert elements[0].text == "News Around NOAA"
 
 
@@ -481,7 +512,7 @@ def test_auto_partition_works_with_unstructured_jsons_from_file():
 @pytest.mark.skipif(odt_not_supported, reason="odt not supported in this version of pypandoc.")
 def test_auto_partition_odt_from_filename():
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake.odt")
-    elements = partition(filename=filename)
+    elements = partition(filename=filename, strategy="hi_res")
     assert elements == [Title("Lorem ipsum dolor sit amet.")]
 
 
@@ -490,7 +521,7 @@ def test_auto_partition_odt_from_filename():
 def test_auto_partition_odt_from_file():
     filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake.odt")
     with open(filename, "rb") as f:
-        elements = partition(file=f)
+        elements = partition(file=f, strategy="hi_res")
 
     assert elements == [Title("Lorem ipsum dolor sit amet.")]